IOT_Backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
paho-mqtt
firebase-admin
requests
